Interfaces corrigidas e otimizadas e classes implementadas. Falta a lógica de atualização dos parâmetros.

+++ src/src/game/view/IUpdateBar.java
@@ -2,7 +2,7 @@ package game.view;
 
 public interface IUpdateBar {
     void setMoney(int quantity);
-    void setApproval(double percentage);
+    void setApproval(int percentage);
     void setDeaths(int quantity);
     void setInfected(int quantity);
     void setImmunized(int quantity);
